# swagger.json
{"openapi":"3.0.0","paths":{"/v1/auth/login":{"post":{"operationId":"AuthController_login","summary":"Login and get JWT token","description":"Authenticate user with email and password to receive a JWT token.\n\n**Demo Users (created by seed script):**\n- `admin@example.com` (admin) - Full access\n- `user1@example.com` (user) - Standard user\n- `user2@example.com` (user) - Standard user\n- `support@example.com` (support) - Read-only access\n\n**Default Password:** All demo users: `password123`\n\n**Note:** Users without passwords can login with the default password.","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/LoginDto"}}}},"responses":{"200":{"description":"Login successful","content":{"application/json":{"schema":{"example":{"access_token":"eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9...","user":{"id":"507f1f77bcf86cd799439011","email":"admin@example.com","role":"admin"}}}}}},"401":{"description":"Invalid credentials"}},"tags":["Authentication"]}},"/v1/docs":{"post":{"operationId":"DocumentsController_uploadDocument","summary":"Upload a document","description":"Upload a document with a primary tag (required) and optional secondary tags. The primary tag creates a folder if it does not exist. Each document must have exactly one primary tag.","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UploadDocumentDto"}}}},"responses":{"201":{"description":"Document uploaded successfully","content":{"application/json":{"schema":{"example":{"_id":"507f1f77bcf86cd799439011","ownerId":"507f1f77bcf86cd799439012","filename":"invoice-2025-01.pdf","mime":"application/pdf","textContent":"Invoice content...","createdAt":"2025-01-13T10:00:00.000Z"}}}}},"400":{"description":"Bad request - validation failed"},"401":{"description":"Unauthorized - invalid or missing JWT token"},"403":{"description":"Forbidden - insufficient permissions"}},"tags":["Documents"],"security":[{"JWT-auth":[]}]}},"/v1/folders":{"get":{"operationId":"FoldersController_getFolders","summary":"List all folders","description":"Returns all primary-tag folders with document counts. Folders are automatically created when documents are uploaded with primary tags.","parameters":[],"responses":{"200":{"description":"List of folders with document counts","content":{"application/json":{"schema":{"type":"array","items":{"type":"object","properties":{"id":{"type":"string","example":"507f1f77bcf86cd799439011"},"name":{"type":"string","example":"invoices-2025"},"count":{"type":"number","example":5}}},"example":[{"id":"507f1f77bcf86cd799439011","name":"invoices-2025","count":5},{"id":"507f1f77bcf86cd799439012","name":"legal","count":2},{"id":"507f1f77bcf86cd799439013","name":"receipts","count":3}]}}}},"401":{"description":"Unauthorized"}},"tags":["Folders"],"security":[{"JWT-auth":[]}]}},"/v1/folders/{tag}/docs":{"get":{"operationId":"FoldersController_getFolderDocuments","summary":"Get documents in a folder","description":"Returns all documents where the specified tag is the primary tag. The tag parameter can be either the tag ID or tag name. Only returns documents owned by the authenticated user.","parameters":[{"name":"tag","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":"List of documents in the folder","content":{"application/json":{"schema":{"type":"array","items":{"type":"object","properties":{"_id":{"type":"string"},"ownerId":{"type":"string"},"filename":{"type":"string"},"mime":{"type":"string"},"textContent":{"type":"string"},"createdAt":{"type":"string","format":"date-time"}}}}}}},"401":{"description":"Unauthorized"},"404":{"description":"Folder (tag) not found"}},"tags":["Folders"],"security":[{"JWT-auth":[]}]}},"/v1/search":{"get":{"operationId":"SearchController_searchDocuments","summary":"Search documents","description":"Full-text search across documents with optional scope filtering.\n\n**Scope Rules:**\n- Use `scope=folder` to search within all folders (documents with primary tags)\n- Use `scope=files` with `ids[]` to search within specific document IDs\n- **Cannot use both folder scope and file IDs together**\n\n**Examples:**\n- Search all: `/v1/search?q=invoice`\n- Search in folders: `/v1/search?q=invoice&scope=folder`\n- Search specific files: `/v1/search?q=invoice&scope=files&ids[]=doc1&ids[]=doc2`","parameters":[{"name":"q","required":true,"in":"query","description":"Search query string","schema":{"example":"invoice","type":"string"}},{"name":"scope","required":false,"in":"query","description":"Search scope","schema":{"enum":["folder","files"],"type":"string"}},{"name":"ids","required":false,"in":"query","description":"Document IDs (required when scope=files)","schema":{"example":["doc1","doc2"],"type":"array","items":{"type":"string"}}}],"responses":{"200":{"description":"List of matching documents","content":{"application/json":{"schema":{"type":"array","items":{"type":"object","properties":{"_id":{"type":"string"},"filename":{"type":"string"},"textContent":{"type":"string"}}}}}}},"400":{"description":"Bad request - invalid scope combination"},"401":{"description":"Unauthorized"}},"tags":["Search"],"security":[{"JWT-auth":[]}]}},"/v1/actions/run":{"post":{"operationId":"ActionsController_runAction","summary":"Run scoped action on documents","description":"Execute actions on a set of documents defined by scope (folder or files).\n\n**Scope Types:**\n- `folder`: Process all documents in a folder (primary tag)\n- `files`: Process specific documents by IDs\n\n**Actions:**\n- `make_document`: Creates a new document with generated text\n- `make_csv`: Generates a CSV file and stores as a document\n\n**Usage:**\n- Each request consumes **5 credits**\n- Credits are tracked in the Usage collection\n\n**Validation:**\n- Cannot use both folder scope and file IDs together\n- Folder scope requires `name` parameter\n- Files scope requires `ids` array","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/RunActionDto"},"examples":{"folderScope":{"summary":"Folder scope example","value":{"scope":{"type":"folder","name":"invoices-2025"},"messages":[{"role":"user","content":"make a CSV of vendor totals"}],"actions":["make_document","make_csv"]}},"filesScope":{"summary":"Files scope example","value":{"scope":{"type":"files","ids":["doc1","doc2"]},"messages":[{"role":"user","content":"summarize these documents"}],"actions":["make_document"]}}}}}},"responses":{"200":{"description":"Action executed successfully","content":{"application/json":{"schema":{"example":{"success":true,"createdDocuments":[{"_id":"doc1","filename":"generated-123.txt","mime":"text/plain"},{"_id":"doc2","filename":"generated-123.csv","mime":"text/csv"}],"creditsUsed":5}}}}},"400":{"description":"Bad request - invalid scope or validation failed"},"401":{"description":"Unauthorized"},"403":{"description":"Forbidden - insufficient permissions"},"404":{"description":"No documents found in scope"}},"tags":["Actions"],"security":[{"JWT-auth":[]}]}},"/v1/actions/usage/month":{"get":{"operationId":"ActionsController_getUsageForMonth","summary":"Get usage statistics for a month","description":"Returns the total credits consumed by the authenticated user for the specified month. Each action request consumes 5 credits.","parameters":[{"name":"year","required":false,"in":"query","description":"Year (defaults to current year)","schema":{"example":2025,"type":"number"}},{"name":"month","required":false,"in":"query","description":"Month 1-12 (defaults to current month)","schema":{"example":11,"type":"number"}}],"responses":{"200":{"description":"Usage statistics","content":{"application/json":{"schema":{"example":{"userId":"507f1f77bcf86cd799439012","credits":15,"year":2025,"month":11}}}}},"401":{"description":"Unauthorized"},"403":{"description":"Forbidden"}},"tags":["Actions"],"security":[{"JWT-auth":[]}]}},"/v1/webhooks/ocr":{"post":{"operationId":"WebhooksController_processOcrWebhook","summary":"Process OCR webhook","description":"Ingest OCR text content and classify it as \"official\" or \"ad\".\n\n**Classification Rules:**\n- **Official**: Contains financial/legal terms (contract, invoice, agreement, legal, etc.)\n- **Ad**: Contains promotional terms (sale, discount, limited time, unsubscribe, etc.)\n\n**For Ads:**\n- Extracts unsubscribe information (email/URL)\n- Creates a Task with status \"pending\"\n- Rate limit: Maximum 3 tasks per sender (source) per day per user\n- If rate limit exceeded, returns processed: false with reason\n\n**Audit:**\n- All webhook events are logged in the audit log","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/OcrWebhookDto"},"examples":{"ad":{"summary":"Ad with unsubscribe","value":{"source":"scanner-01","imageId":"img_123","text":"LIMITED TIME SALE! Get 50% off. unsubscribe: mailto:stop@brand.com","meta":{"address":"123 Main St"}}},"official":{"summary":"Official document","value":{"source":"scanner-01","imageId":"img_124","text":"Service Agreement. Terms and conditions apply. Legal document with invoice details.","meta":{}}}}}}},"responses":{"200":{"description":"Webhook processed successfully","content":{"application/json":{"schema":{"example":{"imageId":"img_123","classification":"ad","processed":true,"taskId":"507f1f77bcf86cd799439011"}}}}},"401":{"description":"Unauthorized"}},"tags":["Webhooks"],"security":[{"JWT-auth":[]}]}},"/v1/metrics":{"get":{"operationId":"MetricsController_getMetrics","summary":"Get system metrics","description":"Returns system-wide metrics. Requires admin, support, or moderator role.","parameters":[],"responses":{"200":{"description":"System metrics","content":{"application/json":{"schema":{"example":{"docs_total":123,"folders_total":7,"actions_month":42,"tasks_today":5}}}}},"401":{"description":"Unauthorized"},"403":{"description":"Forbidden - requires admin/support/moderator role"}},"tags":["Metrics"],"security":[{"JWT-auth":[]}]}}},"info":{"title":"Skylad Backend API","description":"Document management API with folders, tags, OCR webhooks, and RBAC.\n\n## Features\n- Document management with primary/secondary tags\n- Scoped actions for document processing\n- OCR webhook ingestion with classification\n- Role-based access control (RBAC)\n- Comprehensive auditing and metrics\n\n## Demo Users\n\nThe following users are created by the seed script (`npm run seed`). All users have the default password: **`password123`**\n\n| Email | Role | Description |\n|-------|------|-------------|\n| `admin@example.com` | admin | Full access to all resources and metrics |\n| `user1@example.com` | user | Standard user with CRUD on own resources |\n| `user2@example.com` | user | Standard user with CRUD on own resources |\n| `support@example.com` | support | Read-only access to all resources |\n\n**Getting Started:**\n1. Run `npm run seed` to create demo users\n2. Login via `POST /v1/auth/login` with any user's email and password `password123`\n3. Use the returned `access_token` in the `Authorization: Bearer <token>` header\n\n## Authentication\nAll endpoints require JWT Bearer token authentication. Use the login endpoint or seed script to get tokens.\n\n## API Flows\n\n### Document Management Flow\n1. Upload document with primary tag → Creates folder automatically\n2. List folders → See all primary-tag folders with counts\n3. Get folder documents → Retrieve documents in a specific folder\n4. Search documents → Full-text search with scope filtering\n\n### Scoped Actions Flow\n1. Run action with folder/files scope → Processes documents\n2. Generates derived files (documents/CSV) → Creates new documents\n3. Tracks usage → 5 credits per request\n4. View usage → Get monthly credit totals\n\n### OCR Webhook Flow\n1. Post OCR webhook → Classifies content (official/ad)\n2. For ads → Extracts unsubscribe info\n3. Creates task → With rate limiting (3 per sender/day)\n4. Logs event → Audit trail maintained\n\n### RBAC Flow\n- **User**: CRUD on own docs/tags, run actions, view own usage\n- **Support/Moderator**: Read-only access to all resources\n- **Admin**: Full access to all resources including metrics\n","version":"1.0.0","contact":{}},"tags":[{"name":"Authentication","description":"User authentication and JWT token generation"},{"name":"Documents","description":"Document management with folders and tags"},{"name":"Folders","description":"Folder operations (primary-tag based)"},{"name":"Search","description":"Full-text search across documents"},{"name":"Actions","description":"Scoped actions for document processing"},{"name":"Webhooks","description":"OCR webhook ingestion and classification"},{"name":"Metrics","description":"System metrics and statistics"}],"servers":[],"components":{"securitySchemes":{"JWT-auth":{"scheme":"bearer","bearerFormat":"JWT","type":"http","name":"JWT","description":"Enter JWT token","in":"header"}},"schemas":{"LoginDto":{"type":"object","properties":{"email":{"type":"string","description":"User email address","example":"admin@example.com"},"password":{"type":"string","description":"User password","example":"password123","writeOnly":true}},"required":["email","password"]},"UploadDocumentDto":{"type":"object","properties":{"filename":{"type":"string","description":"Name of the file","example":"invoice-2025-01.pdf"},"mime":{"type":"string","description":"MIME type of the file","example":"application/pdf"},"textContent":{"type":"string","description":"Text content extracted from the document","example":"Invoice for January 2025. Total: $1,500.00. Payment due: 30 days."},"primaryTag":{"type":"string","description":"Primary tag name (creates folder if not exists). Each document must have exactly one primary tag.","example":"invoices-2025"},"secondaryTags":{"description":"Optional array of secondary tag names","example":["january","billing"],"type":"array","items":{"type":"string"}}},"required":["filename","mime","textContent","primaryTag"]},"ScopeDto":{"type":"object","properties":{"type":{"type":"string","description":"Scope type: \"folder\" to process all documents in a folder, \"files\" to process specific documents","enum":["folder","files"],"example":"folder"},"name":{"type":"string","description":"Folder name (required when type=folder). Must be a primary tag name.","example":"invoices-2025"},"ids":{"description":"Array of document IDs (required when type=files). Cannot be used with type=folder.","example":["doc1","doc2"],"type":"array","items":{"type":"string"}}},"required":["type"]},"MessageDto":{"type":"object","properties":{"role":{"type":"string","description":"Message role","example":"user"},"content":{"type":"string","description":"Message content/instruction","example":"make a CSV of vendor totals"}},"required":["role","content"]},"RunActionDto":{"type":"object","properties":{"scope":{"description":"Scope definition - either folder or files, not both","allOf":[{"$ref":"#/components/schemas/ScopeDto"}]},"messages":{"description":"Array of messages/instructions for the action","example":[{"role":"user","content":"make a CSV of vendor totals"}],"type":"array","items":{"$ref":"#/components/schemas/MessageDto"}},"actions":{"description":"Array of actions to execute. Supported: \"make_document\", \"make_csv\"","example":["make_document","make_csv"],"type":"array","items":{"type":"string"}}},"required":["scope","messages","actions"]},"OcrWebhookDto":{"type":"object","properties":{"source":{"type":"string","description":"Source identifier (e.g., scanner ID). Used for rate limiting (max 3 tasks per source per day per user)","example":"scanner-01"},"imageId":{"type":"string","description":"Unique identifier for the OCR image","example":"img_123"},"text":{"type":"string","description":"Text content extracted from OCR. Used for classification (official/ad)","example":"LIMITED TIME SALE! Get 50% off. unsubscribe: mailto:stop@brand.com"},"meta":{"type":"object","description":"Additional metadata","example":{"address":"123 Main St"}}},"required":["source","imageId","text"]}}}}